MB-2964 Reorganize utils
