+++ utils/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-""" This package is for utility code that is independent from the Locust codebase """
+""" This package is for utility code that is independent from the Locust codebase. """
